Put DPBasedJoinPlanOptimizer into package se.liu.ida.hefquin.engine.queryproc.impl.optimizer.simple; Move two helper methods into DPBasedJoinPlanOptimizer
