+++ src/main/java/se/liu/ida/hefquin/engine/queryproc/impl/optimizer/simple/DPBasedJoinPlanOptimizer.java
@@ -1,4 +1,4 @@
-package se.liu.ida.hefquin.engine.queryproc.impl.optimizer.DynamicProgramming;
+package se.liu.ida.hefquin.engine.queryproc.impl.optimizer.simple;
 
 import se.liu.ida.hefquin.engine.queryplan.PhysicalPlan;
 import se.liu.ida.hefquin.engine.queryplan.utils.PhysicalPlanFactory;
@@ -6,7 +6,6 @@ import se.liu.ida.hefquin.engine.queryproc.QueryOptimizationException;
 import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.QueryOptimizationContext;
 import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.evolutionaryAlgorithm.PhysicalPlanWithCost;
 import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.evolutionaryAlgorithm.PhysicalPlanWithCostUtils;
-import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.simple.JoinPlanOptimizerBase;
 
 import java.util.*;
 
@@ -42,13 +41,13 @@ public class DPBasedJoinPlanOptimizer extends JoinPlanOptimizerBase {
 
             for ( int num = 2; num < subplans.size(); num ++ ){
                 // Get all subsets with size num.
-                final List<List<PhysicalPlan>> subsets = DPBasedOptimizerHelper.getSubSet(subplans, num);
+                final List<List<PhysicalPlan>> subsets = getSubSet(subplans, num);
 
                 for( final List<PhysicalPlan> plans : subsets ){
                     final List<PhysicalPlan> candidatePlans = new ArrayList<>();
 
                     // Split the current set of subplans into two subsets, and create candidate plans with join for each of the combinations.
-                    DPBasedOptimizerHelper.splitIntoSubSets(plans).forEach(l->{
+                    splitIntoSubSets(plans).forEach(l->{
                         final PhysicalPlan newPlan = PhysicalPlanFactory.createPlanWithJoin( optPlan.get(l.get(0)), optPlan.get(l.get(1)));
                         candidatePlans.add(newPlan);
                     });
@@ -62,6 +61,68 @@ public class DPBasedJoinPlanOptimizer extends JoinPlanOptimizerBase {
             return optPlan.get( subplans );
         }
 
+    }
+
+    // This method returns all subsets (with the given size) of the given superset.
+    public static <T> List<List<T>> getSubSet( final List<T> superset, final int n ){
+        final List<List<T>> result = new ArrayList<>();
+        if( n==0 ){
+            result.add( new ArrayList<>() );
+            return result;
+        }
+
+        final List<List<T>> tempList = new ArrayList<>();
+        tempList.add( new ArrayList<>() );
+
+        for ( T element : superset ) {
+            final int size = tempList.size();
+            for ( int j = 0; j < size; j++ ){
+                // Stop adding more elements to the subset if its size is more than n.
+                if( tempList.get(j).size() >= n ){
+                    continue;
+                }
+
+                // Create a copy of the current subsets, then update the copy by adding an another element.
+                final List<T> clone = new ArrayList<>( tempList.get(j) );
+                clone.add( element );
+                tempList.add(clone);
+
+                // Add the subset to the result if its size is n.
+                if( clone.size() == n ){
+                    result.add(clone);
+                }
+            }
+        }
+
+        return result;
+    }
+
+    // Split a superset into two subsets. This method returns all possible combinations of subsets.
+    public static <T> List<List<List<T>>> splitIntoSubSets( final List<T> superset ){
+        final List<List<T>> left = new ArrayList<>();
+        final List<List<T>> right = new ArrayList<>();
+        final List<List<List<T>>> result = new ArrayList<>();
+
+        left.add( new ArrayList<>() );
+        right.add( new ArrayList<>(superset) );
+
+        for ( T element : superset ) {
+            final int leftSize = left.size();
+            for ( int j = 0; j < leftSize; j++ ){
+                final List<T> leftClone = new ArrayList<>( left.get(j) );
+                leftClone.add( element );
+                left.add( leftClone );
+
+                final List<T> rightClone = new ArrayList<>( right.get(j) );
+                rightClone.remove( element );
+                right.add( rightClone );
+
+                if ( leftClone.size() != 0 && rightClone.size() != 0 )
+                    result.add( Arrays.asList( leftClone, rightClone ) );
+
+            }
+        }
+        return result;
     }
 
 }
